169-5 ShareLog to models
