+++ nomium/shares-logger/src/models/mod.rs
@@ -1,0 +1,2 @@
+mod share;
+pub use share::ShareLog;
